fix(web): stop importing visual workflow demo from fixture
Vercel ignores *.fixture.ts, so next build cannot resolve the sample
graph the editor loaded. Keep the draft in the mock module instead.
